requirements.txt
anyio==4.4.0
asgiref==3.8.1
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
Django==3.2.25
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
h11==0.14.0
harness-featureflags==1.6.1
httpcore==1.0.5
httpx==0.27.0
idna==3.7
mmh3==4.1.0
pillow==10.3.0
PyJWT==2.8.0
pytz==2024.1
requests==2.32.3
sniffio==1.3.1
sqlparse==0.5.0
tenacity==8.2.2
urllib3==2.2.1
requests>=2.25.1